refactor(reports): split route into per-report content components
`apps/webapp/app/routes/_layout+/reports.$reportId.tsx` had grown to
2832 lines, mixing the route's loader/dispatch with 9 inline page-sized
Content components, 5 module-scope `_COLUMNS` constants, and a handful
of small UI helpers and formatter functions. This violated the existing
shelf pattern of a thin route + components living under
`~/components/<feature>/`. The `~/components/reports/` folder already
contained 20 extracted pieces (charts, KPI grid, hero, donuts, table,
pagination, etc.) with kebab-case filenames and a barrel `index.ts` —
the Content components belonged there too.

This change is structurally byte-equivalent. No behavior changes; the
route file's loader, ReportPage, and dispatch helpers are unchanged.
Each Content component is a verbatim move into its own file with the
imports adjusted to point at the right modules.

Extracted (12 new files in `apps/webapp/app/components/reports/`):

  Per-report Content components:
  - booking-compliance-content.tsx (270 lines, owns formatStatus,
    getStatusVariant, formatLateness as private helpers)
  - overdue-items-content.tsx (220 lines)
  - idle-assets-content.tsx (221 lines, owns IDLE_ASSETS_COLUMNS)
  - custody-snapshot-content.tsx (222 lines, columns inline because
    they close over `maxDays` for the relative tenure-bar width)
  - top-booked-assets-content.tsx (303 lines, owns
    TOP_BOOKED_ASSETS_COLUMNS and private AvgDurationHeader)
  - asset-distribution-content.tsx (126 lines)
  - asset-inventory-content.tsx (233 lines, owns ASSET_INVENTORY_COLUMNS,
    formatAssetStatus, getAssetStatusVariant)
  - monthly-booking-trends-content.tsx (241 lines)
  - asset-utilization-content.tsx (215 lines, owns
    ASSET_UTILIZATION_COLUMNS)
  - asset-activity-content.tsx (231 lines, owns ASSET_ACTIVITY_COLUMNS)

  Filter UI helpers:
  - timeframe-range-indicator.tsx (74 lines)
  - idle-threshold-selector.tsx (69 lines)

The `~/components/reports/index.ts` barrel adds re-exports for all 12
new modules under appropriate grouping comments.

Route file now: 836 lines (down from 2832, -2014 LOC). It contains
purely loader + ReportPage + 5 dispatch predicate helpers
(getEmptyStateTitle, getEmptyStateDescription, showTimeframePicker,
getExcludedPresets, getEmptyStateCta).

Each new file follows the conventions used by sibling files in the same
folder: file-level `@file` JSDoc with `@see` back to the route, inline
`type Props = {...}` with per-prop JSDoc, named export, row types
imported from `~/modules/reports/types`, cell types from
`~/components/reports/report-table`. Module-scope `_COLUMNS` constants
are preserved with their existing JSDoc explaining why hoisting matters
(stable cell function refs prevent the AssetCell/AssetImage remount
storms documented in earlier fixes).

Verified: typecheck clean, lint at warning baseline (no new warnings),
all 1777 unit tests pass via `pnpm webapp:validate`.

Manual browser verification still recommended before merge: navigate
through every `/reports/*` URL and confirm there's no flood of
NS_BINDING_ABORTED lazy-img requests in DevTools (regression check for
the prior re-render-storm fix).
