+++ apps/webapp/app/routes/_layout+/reports.$reportId.tsx
@@ -2,44 +2,47 @@
  * Report Runner Route
  *
  * Dynamic route that renders a specific report based on the reportId param.
- * Handles loading data, applying filters, and rendering the appropriate report.
+ * The route owns three concerns:
+ *   1. Permission + data loading via the per-report `*Report` server helpers.
+ *   2. The page chrome (header, export buttons, timeframe filter bar, footer).
+ *   3. Dispatch to the right Content component based on `reportId`.
+ *
+ * Each report's UI lives in its own `*-content.tsx` file under
+ * `~/components/reports/` (one component per file, matching the existing
+ * shelf convention for that folder). This keeps the route file thin and
+ * lets each report evolve independently.
  *
  * @see {@link file://../../modules/reports/registry.ts}
  * @see {@link file://../../modules/reports/helpers.server.ts}
+ * @see {@link file://../../components/reports/index.ts}
  */
 
-import { useCallback, useMemo, useState } from "react";
-import type { ColumnDef } from "@tanstack/react-table";
+import { useCallback, useState } from "react";
 import { useSetAtom } from "jotai";
 import type { LoaderFunctionArgs, MetaFunction } from "react-router";
-import {
-  data,
-  Link,
-  useLoaderData,
-  useNavigate,
-  useNavigation,
-} from "react-router";
+import { data, useLoaderData, useNavigate, useNavigation } from "react-router";
 
 import { showNotificationAtom } from "~/atoms/notifications";
-import { AssetImage } from "~/components/assets/asset-image";
 import Header from "~/components/layout/header";
 import {
-  ReportFooter,
-  TimeframePicker,
-  ReportTable,
-  StatusCell,
-  DateCell,
-  NumberCell,
-  AssetCell,
+  AssetActivityContent,
+  AssetDistributionContent,
+  AssetInventoryContent,
+  AssetUtilizationContent,
+  BookingComplianceContent,
+  CustodySnapshotContent,
+  IdleAssetsContent,
+  IdleThresholdSelector,
+  MonthlyBookingTrendsContent,
+  OverdueItemsContent,
   ReportEmptyState,
-  ComplianceHero,
+  ReportFooter,
   ReportPdf,
-  BarChart,
-  ChartCard,
-  DistributionDonut,
+  TimeframePicker,
+  TimeframeRangeIndicator,
+  TopBookedAssetsContent,
 } from "~/components/reports";
 import { Button } from "~/components/shared/button";
-import { InfoTooltip } from "~/components/shared/info-tooltip";
 import { useSearchParams } from "~/hooks/search-params";
 import {
   resolveTimeframe,
@@ -57,21 +60,19 @@ import {
 } from "~/modules/reports/helpers.server";
 import { getReportById } from "~/modules/reports/registry";
 import type {
-  BookingComplianceRow,
-  OverdueItemRow,
-  IdleAssetRow,
-  CustodySnapshotRow,
-  TopBookedAssetRow,
-  AssetInventoryRow,
-  MonthlyBookingTrendRow,
-  AssetUtilizationRow,
   AssetActivityRow,
-  DistributionBreakdown,
+  AssetInventoryRow,
+  AssetUtilizationRow,
+  BookingComplianceRow,
   ChartSeries,
-  ReportKpi,
+  CustodySnapshotRow,
+  DistributionBreakdown,
+  IdleAssetRow,
+  MonthlyBookingTrendRow,
+  OverdueItemRow,
   ReportPayload,
   TimeframePreset,
-  ResolvedTimeframe,
+  TopBookedAssetRow,
 } from "~/modules/reports/types";
 import { appendToMetaTitle } from "~/utils/append-to-meta-title";
 import { ShelfError } from "~/utils/error";
@@ -80,7 +81,6 @@ import {
   PermissionEntity,
 } from "~/utils/permissions/permission.data";
 import { requirePermission } from "~/utils/roles.server";
-
 import { tw } from "~/utils/tw";
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => [
@@ -299,6 +299,7 @@ export default function ReportPage() {
   const [isExporting, setIsExporting] = useState(false);
   const showNotification = useSetAtom(showNotificationAtom);
   const navigation = useNavigation();
+  const navigate = useNavigate();
 
   // Show loading state when navigating (timeframe change, pagination, etc.)
   const isLoading = navigation.state === "loading";
@@ -390,8 +391,6 @@ export default function ReportPage() {
       setIsExporting(false);
     }
   };
-
-  const navigate = useNavigate();
 
   // Row-click handlers are wrapped in `useCallback` so their identity is stable
   // across re-renders. Without this, every ReportPage re-render produced a new
@@ -709,364 +708,6 @@ export default function ReportPage() {
 }
 
 // -----------------------------------------------------------------------------
-// Filter components
-// -----------------------------------------------------------------------------
-
-/**
- * Shows the actual date range and day count for the selected timeframe.
- * Helps users understand why "This quarter" might show fewer bookings than "Last 90 days"
- * (e.g., Q2 starting Apr 1 is only 24 days so far vs a full 90 days).
- */
-function TimeframeRangeIndicator({
-  timeframe,
-}: {
-  timeframe: ResolvedTimeframe;
-}) {
-  const { from, to } = timeframe;
-
-  // Calculate day count (inclusive of both start and end dates)
-  // Normalize both dates to midnight to avoid time-of-day issues
-  const fromMidnight = new Date(
-    from.getFullYear(),
-    from.getMonth(),
-    from.getDate()
-  );
-  const toMidnight = new Date(to.getFullYear(), to.getMonth(), to.getDate());
-  const dayCount =
-    Math.round(
-      (toMidnight.getTime() - fromMidnight.getTime()) / (1000 * 60 * 60 * 24)
-    ) + 1;
-
-  // Format dates compactly
-  const formatDate = (date: Date) =>
-    date.toLocaleDateString("en-US", {
-      month: "short",
-      day: "numeric",
-    });
-
-  // Check if dates span different years
-  const fromYear = from.getFullYear();
-  const toYear = to.getFullYear();
-  const showYear = fromYear !== toYear || toYear !== new Date().getFullYear();
-
-  const fromStr =
-    formatDate(from) + (showYear && fromYear !== toYear ? `, ${fromYear}` : "");
-  const toStr = formatDate(to) + (showYear ? `, ${toYear}` : "");
-
-  return (
-    <span className="hidden text-xs text-gray-500 sm:inline-flex sm:items-center sm:gap-1.5">
-      <span>
-        {fromStr} – {toStr}
-      </span>
-      <span className="text-gray-300">·</span>
-      <span className="font-medium text-gray-600">
-        {dayCount} day{dayCount !== 1 ? "s" : ""}
-      </span>
-    </span>
-  );
-}
-
-/**
- * Idle threshold selector for the Idle Assets report.
- * Allows users to define what "idle" means: 30, 60, or 90+ days of inactivity.
- */
-function IdleThresholdSelector({
-  value,
-  onChange,
-  disabled = false,
-}: {
-  value: number;
-  onChange: (days: number) => void;
-  disabled?: boolean;
-}) {
-  const thresholds = [
-    { days: 30, label: "30 days" },
-    { days: 60, label: "60 days" },
-    { days: 90, label: "90 days" },
-  ];
-
-  return (
-    <div className="flex items-center gap-3">
-      <span className="text-sm text-gray-600">Unused for:</span>
-      <div className="flex items-center gap-1 rounded border border-gray-200 bg-white p-1">
-        {thresholds.map((t) => (
-          <button
-            key={t.days}
-            type="button"
-            onClick={() => onChange(t.days)}
-            disabled={disabled}
-            className={tw(
-              "rounded-md px-3 py-1.5 text-sm font-medium transition-colors",
-              "focus:outline-none focus-visible:ring-2 focus-visible:ring-primary-500 focus-visible:ring-offset-1",
-              "disabled:cursor-not-allowed disabled:opacity-50",
-              value === t.days
-                ? "bg-primary-600 text-white"
-                : "text-gray-600 hover:bg-gray-100 hover:text-gray-900"
-            )}
-          >
-            {t.label}
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// Report-specific content components
-// -----------------------------------------------------------------------------
-
-function BookingComplianceContent({
-  rows,
-  complianceData,
-  totalBookings,
-  timeframeLabel,
-  onRowClick,
-}: {
-  rows: BookingComplianceRow[];
-  complianceData?: {
-    onTime: number;
-    late: number;
-    rate: number | null;
-    priorPeriod?: { rate: number; delta: number; periodLabel: string };
-  };
-  totalBookings: number;
-  timeframeLabel?: string;
-  onRowClick?: (row: BookingComplianceRow) => void;
-}) {
-  const [searchParams, setSearchParams] = useSearchParams();
-
-  // Map column accessorKey to server sort column
-  const columnToSortColumn: Record<string, BookingComplianceSortColumn> = {
-    bookingName: "bookingName",
-    status: "status",
-    custodian: "custodian",
-    assetCount: "assetCount",
-    scheduledEnd: "scheduledEnd",
-    latenessMs: "returnStatus",
-  };
-
-  // Read current sort from URL
-  const currentSortBy = searchParams.get("sortBy") || "scheduledEnd";
-  const currentSortOrder = searchParams.get("sortOrder") || "desc";
-
-  // Map server sort column back to column ID for TanStack
-  const sortColumnToAccessor: Record<string, string> = {
-    bookingName: "bookingName",
-    status: "status",
-    custodian: "custodian",
-    assetCount: "assetCount",
-    scheduledEnd: "scheduledEnd",
-    returnStatus: "latenessMs",
-  };
-  const initialSortColumn =
-    sortColumnToAccessor[currentSortBy] || "scheduledEnd";
-
-  // Handle sort change - update URL to trigger server-side sort
-  const handleSortChange = (columnId: string, direction: "asc" | "desc") => {
-    const serverColumn = columnToSortColumn[columnId];
-    if (!serverColumn) return;
-
-    const params = new URLSearchParams(searchParams);
-    params.set("sortBy", serverColumn);
-    params.set("sortOrder", direction);
-    // Reset to page 1 when sort changes
-    params.set("page", "1");
-    setSearchParams(params, { replace: true });
-  };
-
-  // Column definitions for the booking compliance table
-  const columns: ColumnDef<BookingComplianceRow>[] = [
-    {
-      accessorKey: "bookingName",
-      header: "Booking",
-      cell: ({ row }) => (
-        <span className="font-medium">{row.original.bookingName}</span>
-      ),
-    },
-    {
-      accessorKey: "status",
-      header: "Status",
-      cell: ({ row }) => {
-        const status = row.original.status;
-        const variant = getStatusVariant(status);
-        return <StatusCell status={formatStatus(status)} variant={variant} />;
-      },
-    },
-    {
-      accessorKey: "custodian",
-      header: "Booked by",
-      cell: ({ row }) =>
-        row.original.custodian || <span className="text-gray-400">—</span>,
-    },
-    {
-      accessorKey: "assetCount",
-      header: "Assets",
-      cell: ({ row }) => <NumberCell value={row.original.assetCount} />,
-    },
-    {
-      accessorKey: "scheduledEnd",
-      header: "Due Date",
-      cell: ({ row }) => <DateCell date={row.original.scheduledEnd} />,
-    },
-    {
-      accessorKey: "latenessMs",
-      header: "Return Status",
-      cell: ({ row }) => {
-        const { isOnTime, latenessMs } = row.original;
-        if (isOnTime) {
-          return <StatusCell status="On time" variant="success" />;
-        }
-        // Format lateness as human-readable
-        const lateness = formatLateness(latenessMs);
-        return (
-          <span className="text-sm font-medium text-orange-600">
-            {lateness}
-          </span>
-        );
-      },
-    },
-  ];
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* The Answer */}
-      <ComplianceHero
-        rate={complianceData?.rate ?? 0}
-        onTime={complianceData?.onTime ?? 0}
-        late={complianceData?.late ?? 0}
-        priorPeriod={complianceData?.priorPeriod}
-        timeframeLabel={timeframeLabel}
-      />
-
-      {/* Booking details table */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">
-            Booking Details
-          </h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalBookings}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          onRowClick={onRowClick}
-          manualSorting
-          initialSorting={[
-            { id: initialSortColumn, desc: currentSortOrder === "desc" },
-          ]}
-          onSortChange={handleSortChange}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No bookings found"
-              description="No bookings match the current filters."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// Helper functions
-function formatStatus(status: string): string {
-  const labels: Record<string, string> = {
-    DRAFT: "Draft",
-    RESERVED: "Reserved",
-    ONGOING: "Ongoing",
-    OVERDUE: "Overdue",
-    COMPLETE: "Complete",
-    CANCELLED: "Cancelled",
-    ARCHIVED: "Archived",
-  };
-  return labels[status] || status;
-}
-
-function getStatusVariant(
-  status: string
-): "success" | "warning" | "error" | "neutral" {
-  switch (status) {
-    case "COMPLETE":
-      return "success";
-    case "ONGOING":
-    case "RESERVED":
-      return "neutral";
-    case "OVERDUE":
-      return "error";
-    case "CANCELLED":
-    case "DRAFT":
-    case "ARCHIVED":
-    default:
-      return "neutral";
-  }
-}
-
-/**
- * Map asset status to badge variant.
- * Matches the colors used in asset-status-badge.tsx:
- * - AVAILABLE → green (success)
- * - IN_CUSTODY → blue
- * - CHECKED_OUT → violet
- */
-function getAssetStatusVariant(
-  status: string
-): "success" | "blue" | "violet" | "neutral" {
-  switch (status) {
-    case "AVAILABLE":
-      return "success";
-    case "IN_CUSTODY":
-      return "blue";
-    case "CHECKED_OUT":
-      return "violet";
-    default:
-      return "neutral";
-  }
-}
-
-/**
- * Format asset status for display.
- * Matches the labels from asset-status-badge.tsx.
- */
-function formatAssetStatus(status: string): string {
-  const labels: Record<string, string> = {
-    AVAILABLE: "Available",
-    IN_CUSTODY: "In custody",
-    CHECKED_OUT: "Checked out",
-  };
-  return labels[status] || status;
-}
-
-/**
- * Format lateness as human-readable string.
- * Positive ms = late, negative ms = early.
- */
-function formatLateness(ms: number | null): string {
-  if (ms === null) return "—";
-
-  const absMs = Math.abs(ms);
-  const minutes = Math.floor(absMs / (1000 * 60));
-  const hours = Math.floor(absMs / (1000 * 60 * 60));
-  const days = Math.floor(absMs / (1000 * 60 * 60 * 24));
-
-  let value: string;
-  if (days > 0) {
-    const remainingHours = hours % 24;
-    value = remainingHours > 0 ? `${days}d ${remainingHours}h` : `${days}d`;
-  } else if (hours > 0) {
-    const remainingMinutes = minutes % 60;
-    value =
-      remainingMinutes > 0 ? `${hours}h ${remainingMinutes}m` : `${hours}h`;
-  } else {
-    value = `${minutes}m`;
-  }
-
-  return ms > 0 ? `${value} late` : `${value} early`;
-}
-
-// -----------------------------------------------------------------------------
 // Empty state content helpers
 //
 // These provide contextually appropriate messaging for each report type.
@@ -1192,1641 +833,4 @@ function getEmptyStateCta(
   // The appropriate action is to adjust the timeframe, which is
   // already available via the TimeframePicker above.
   return null;
-}
-
-// -----------------------------------------------------------------------------
-// R6: Overdue Items Content
-// -----------------------------------------------------------------------------
-
-function OverdueItemsContent({
-  rows,
-  kpis,
-  totalRows,
-  onRowClick,
-}: {
-  rows: OverdueItemRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  onRowClick?: (row: OverdueItemRow) => void;
-}) {
-  // Column definitions for overdue items table
-  const columns: ColumnDef<OverdueItemRow>[] = [
-    {
-      accessorKey: "bookingName",
-      header: "Booking",
-      cell: ({ row }) => (
-        <span className="font-medium">{row.original.bookingName}</span>
-      ),
-    },
-    {
-      accessorKey: "custodian",
-      header: "Booked by",
-      cell: ({ row }) =>
-        row.original.custodian || <span className="text-gray-400">—</span>,
-    },
-    {
-      accessorKey: "uncheckedCount",
-      header: "Assets",
-      cell: ({ row }) => {
-        const { uncheckedCount, assetCount, checkedInCount } = row.original;
-        const hasPartialReturns = checkedInCount > 0;
-        const progressPercent =
-          assetCount > 0 ? (checkedInCount / assetCount) * 100 : 0;
-
-        return (
-          <div className="flex items-center gap-2">
-            <span className="font-medium tabular-nums">
-              {uncheckedCount}
-              <span className="font-normal text-gray-400"> / {assetCount}</span>
-            </span>
-            {/* Progress bar - shows return progress (green = returned) */}
-            {hasPartialReturns && (
-              <div className="h-2 w-12 overflow-hidden rounded-full bg-gray-200">
-                <div
-                  className="h-full rounded-full bg-green-500"
-                  style={{ width: `${progressPercent}%` }}
-                />
-              </div>
-            )}
-          </div>
-        );
-      },
-    },
-    {
-      accessorKey: "scheduledEnd",
-      header: "Due Date",
-      cell: ({ row }) => <DateCell date={row.original.scheduledEnd} />,
-    },
-    {
-      accessorKey: "daysOverdue",
-      header: "Days Overdue",
-      cell: ({ row }) => {
-        const days = row.original.daysOverdue;
-        return (
-          <span
-            className={tw(
-              "inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-semibold",
-              days > 7
-                ? "bg-red-100 text-red-700"
-                : days > 3
-                ? "bg-orange-100 text-orange-700"
-                : "bg-yellow-100 text-yellow-700"
-            )}
-          >
-            {days} days
-          </span>
-        );
-      },
-    },
-    {
-      accessorKey: "valueAtRisk",
-      header: "Value",
-      cell: ({ row }) =>
-        row.original.valueAtRisk ? (
-          `$${row.original.valueAtRisk.toLocaleString()}`
-        ) : (
-          <span className="text-gray-400">—</span>
-        ),
-    },
-  ];
-
-  // Extract KPI values for hero display
-  const totalOverdue =
-    (kpis.find((k) => k.id === "total_overdue")?.rawValue as number) || 0;
-  const assetsAtRisk =
-    (kpis.find((k) => k.id === "total_assets_at_risk")?.rawValue as number) ||
-    0;
-  const valueAtRisk =
-    (kpis.find((k) => k.id === "total_value_at_risk")?.rawValue as number) || 0;
-  const longestOverdue =
-    (kpis.find((k) => k.id === "longest_overdue")?.rawValue as number) || 0;
-
-  return (
-    <div className="space-y-4">
-      {/* Hero section with key metrics */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric - lead with assets (the scope of the problem) */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span
-                className={tw(
-                  "text-3xl font-semibold",
-                  assetsAtRisk > 0 ? "text-red-600" : "text-green-600"
-                )}
-              >
-                {assetsAtRisk}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                {assetsAtRisk === 1 ? "Asset" : "Assets"} Outstanding
-              </span>
-              {totalOverdue > 0 ? (
-                <span className="text-xs text-gray-500">
-                  across {totalOverdue} overdue booking
-                  {totalOverdue !== 1 ? "s" : ""}
-                </span>
-              ) : (
-                <span className="text-xs text-green-600">
-                  All assets returned on time
-                </span>
-              )}
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          {totalOverdue > 0 && (
-            <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-              <div className="flex flex-col">
-                <span className="text-xs text-gray-500">Total Value</span>
-                <span className="text-lg font-medium text-gray-900">
-                  {valueAtRisk > 0 ? `$${valueAtRisk.toLocaleString()}` : "—"}
-                </span>
-              </div>
-              <div className="flex flex-col">
-                <span className="text-xs text-gray-500">Longest Overdue</span>
-                <span className="text-lg font-medium text-red-600">
-                  {longestOverdue > 0 ? `${longestOverdue} days` : "—"}
-                </span>
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-
-      {/* Data table */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">
-            Overdue Bookings
-          </h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          maxHeight="500px"
-          onRowClick={onRowClick}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No items waiting"
-              description="All items have been returned on schedule."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R4: Idle Assets Content
-// -----------------------------------------------------------------------------
-
-/**
- * Column definitions for the Idle Assets table, defined at module scope so
- * the cell function references stay stable across IdleAssetsContent renders.
- * See the loop bug fix in IdleAssetsContent for context.
- */
-const IDLE_ASSETS_COLUMNS: ColumnDef<IdleAssetRow>[] = [
-  {
-    accessorKey: "assetName",
-    header: "Asset",
-    cell: ({ row }) => (
-      <AssetCell
-        name={row.original.assetName}
-        thumbnailImage={row.original.thumbnailImage}
-        assetId={row.original.assetId}
-      />
-    ),
-  },
-  {
-    accessorKey: "daysSinceLastUse",
-    header: "Unused For",
-    cell: ({ row }) => {
-      const days = row.original.daysSinceLastUse;
-      return (
-        <span
-          className={tw(
-            "inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-semibold",
-            days > 90
-              ? "bg-red-100 text-red-700"
-              : days > 60
-              ? "bg-orange-100 text-orange-700"
-              : "bg-yellow-100 text-yellow-700"
-          )}
-        >
-          {days} days
-        </span>
-      );
-    },
-  },
-  {
-    accessorKey: "lastBookedAt",
-    header: "Last Used",
-    cell: ({ row }) =>
-      row.original.lastBookedAt ? (
-        <DateCell date={row.original.lastBookedAt} />
-      ) : (
-        <span className="text-gray-400">Never</span>
-      ),
-  },
-  {
-    accessorKey: "category",
-    header: "Category",
-    cell: ({ row }) =>
-      row.original.category || <span className="text-gray-400">—</span>,
-  },
-  {
-    accessorKey: "location",
-    header: "Location",
-    cell: ({ row }) =>
-      row.original.location || <span className="text-gray-400">—</span>,
-  },
-  {
-    accessorKey: "valuation",
-    header: "Value",
-    cell: ({ row }) =>
-      row.original.valuation ? (
-        `$${row.original.valuation.toLocaleString()}`
-      ) : (
-        <span className="text-gray-400">—</span>
-      ),
-  },
-];
-
-function IdleAssetsContent({
-  rows,
-  kpis,
-  totalRows,
-  timeframeLabel,
-  onRowClick,
-}: {
-  rows: IdleAssetRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  timeframeLabel?: string;
-  onRowClick?: (row: IdleAssetRow) => void;
-}) {
-  // Column definitions for unused assets table.
-  // Memoized so cell function refs are stable across re-renders — without
-  // this, TanStack `flexRender` hands React a new component type on every
-  // render, which unmounts/remounts every AssetCell (and thereby its
-  // AssetImage child), causing image fetches to abort and re-issue in a
-  // loop. The deps list is empty because the cells only read from `row` —
-  // they don't capture any closure variables from this component.
-  const columns: ColumnDef<IdleAssetRow>[] = useMemo(
-    () => IDLE_ASSETS_COLUMNS,
-    []
-  );
-
-  // Extract KPI values
-  const totalIdle =
-    (kpis.find((k) => k.id === "total_idle")?.rawValue as number) || 0;
-  const idlePercentage =
-    (kpis.find((k) => k.id === "idle_percentage")?.rawValue as number) || 0;
-  const totalIdleValue =
-    (kpis.find((k) => k.id === "total_idle_value")?.rawValue as number) || 0;
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span
-                className={tw(
-                  "text-3xl font-semibold",
-                  totalIdle > 10
-                    ? "text-orange-600"
-                    : totalIdle > 0
-                    ? "text-yellow-600"
-                    : "text-green-600"
-                )}
-              >
-                {totalIdle}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                Unused Assets
-              </span>
-              <span className="text-xs text-gray-500">
-                {idlePercentage}% of inventory
-              </span>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Total Value</span>
-              <span className="text-lg font-medium text-gray-900">
-                {totalIdleValue > 0
-                  ? `$${totalIdleValue.toLocaleString()}`
-                  : "—"}
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Data table */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">Unused Assets</h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          onRowClick={onRowClick}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No idle assets"
-              description={`All your assets have been used within the selected threshold${
-                timeframeLabel ? ` (${timeframeLabel.toLowerCase()})` : ""
-              }.`}
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R5: Custody Snapshot Content
-// -----------------------------------------------------------------------------
-
-function CustodySnapshotContent({
-  rows,
-  kpis,
-  totalRows,
-  onRowClick,
-}: {
-  rows: CustodySnapshotRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  onRowClick?: (row: CustodySnapshotRow) => void;
-}) {
-  // Calculate max days for relative bar width
-  const maxDays = Math.max(...rows.map((r) => r.daysInCustody), 1);
-
-  // Column definitions for custody snapshot table.
-  // Memoized so cell function refs are stable across re-renders. Without
-  // this, TanStack flexRender hands React a new component type on every
-  // render → every AssetCell unmounts/remounts → every AssetImage
-  // remounts → image-fetch storm. Deps include `maxDays` because the
-  // tenure-bar cell closes over it.
-  const columns: ColumnDef<CustodySnapshotRow>[] = useMemo(
-    () => [
-      {
-        accessorKey: "assetName",
-        header: "Asset",
-        cell: ({ row }) => (
-          <AssetCell
-            name={row.original.assetName}
-            thumbnailImage={row.original.thumbnailImage}
-            assetId={row.original.assetId}
-          />
-        ),
-      },
-      {
-        accessorKey: "custodianName",
-        header: "Assigned to",
-        cell: ({ row }) => row.original.custodianName,
-      },
-      {
-        accessorKey: "daysInCustody",
-        header: "Days Held",
-        cell: ({ row }) => {
-          const days = row.original.daysInCustody;
-          const percentage = Math.min((days / maxDays) * 100, 100);
-          return (
-            <div className="flex items-center gap-3">
-              {/* Tenure bar - visual indicator of relative duration */}
-              <div className="relative h-2 w-16 overflow-hidden rounded-full bg-gray-100">
-                <div
-                  className="absolute inset-y-0 left-0 rounded-full bg-primary-500 transition-all"
-                  style={{ width: `${percentage}%` }}
-                />
-              </div>
-              {/* Days value */}
-              <span className="min-w-16 text-sm font-medium tabular-nums text-gray-900">
-                {days} <span className="font-normal text-gray-500">days</span>
-              </span>
-            </div>
-          );
-        },
-      },
-      {
-        accessorKey: "assignedAt",
-        header: "Assigned",
-        cell: ({ row }) => <DateCell date={row.original.assignedAt} />,
-      },
-      {
-        accessorKey: "category",
-        header: "Category",
-        cell: ({ row }) =>
-          row.original.category || <span className="text-gray-400">—</span>,
-      },
-      {
-        accessorKey: "location",
-        header: "Location",
-        cell: ({ row }) =>
-          row.original.location || <span className="text-gray-400">—</span>,
-      },
-      {
-        accessorKey: "valuation",
-        header: "Value",
-        cell: ({ row }) =>
-          row.original.valuation ? (
-            `$${row.original.valuation.toLocaleString()}`
-          ) : (
-            <span className="text-gray-400">—</span>
-          ),
-      },
-    ],
-    [maxDays]
-  );
-
-  // Extract KPI values
-  const totalInCustody =
-    (kpis.find((k) => k.id === "total_in_custody")?.rawValue as number) || 0;
-  const totalCustodians =
-    (kpis.find((k) => k.id === "total_custodians")?.rawValue as number) || 0;
-  const totalCustodyValue =
-    (kpis.find((k) => k.id === "total_custody_value")?.rawValue as number) || 0;
-  const avgDaysInCustody =
-    (kpis.find((k) => k.id === "avg_days_in_custody")?.rawValue as number) || 0;
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span className="text-3xl font-semibold text-gray-900">
-                {totalInCustody}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                Assets Currently Assigned
-              </span>
-              <span className="text-xs text-gray-500">
-                Across {totalCustodians} team member
-                {totalCustodians !== 1 ? "s" : ""}
-              </span>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Total Value</span>
-              <span className="text-lg font-medium text-gray-900">
-                {totalCustodyValue > 0
-                  ? `$${totalCustodyValue.toLocaleString()}`
-                  : "—"}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Avg. Tenure</span>
-              <span className="text-lg font-medium text-gray-900">
-                {avgDaysInCustody > 0
-                  ? `${Math.round(avgDaysInCustody)} days`
-                  : "—"}
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Data table */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">
-            Current Assignments
-          </h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          onRowClick={onRowClick}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No assets assigned"
-              description="No assets are currently assigned to team members."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R3: Top Booked Assets Content
-// -----------------------------------------------------------------------------
-
-// why: defined at module scope so the function identity is stable across
-// renders. TanStack's flexRender treats a different `header` function ref as a
-// different component type and remounts the whole subtree — for a header that
-// renders Radix's InfoTooltip, the ref-attach/detach churn triggers
-// setTrigger(...) loops during navigation transitions ("Maximum update depth
-// exceeded" — see SHELF-WEBAPP-1J4).
-function AvgDurationHeader() {
-  return (
-    <span className="flex items-center gap-1">
-      Avg Duration
-      <InfoTooltip
-        iconClassName="size-3.5"
-        content={
-          <p>
-            <strong>Average booking duration</strong> — How long this asset is
-            typically kept per booking. Calculated as total days booked ÷ number
-            of bookings.
-          </p>
-        }
-      />
-    </span>
-  );
-}
-
-/**
- * Column definitions for the Top Booked Assets table, defined at module
- * scope so cell function refs are stable across TopBookedAssetsContent
- * renders. See the loop fix explanation in IDLE_ASSETS_COLUMNS / the
- * commit history of this file.
- */
-const TOP_BOOKED_ASSETS_COLUMNS: ColumnDef<TopBookedAssetRow>[] = [
-  {
-    accessorKey: "assetName",
-    header: "Asset",
-    cell: ({ row }) => (
-      <AssetCell
-        name={row.original.assetName}
-        thumbnailImage={row.original.thumbnailImage}
-        assetId={row.original.assetId}
-      />
-    ),
-  },
-  {
-    accessorKey: "bookingCount",
-    header: "Bookings",
-    cell: ({ row }) => (
-      <Link
-        to={`/assets/${row.original.assetId}/bookings`}
-        onClick={(e) => e.stopPropagation()}
-        className="inline-flex items-center gap-1 rounded-full bg-blue-100 px-2 py-0.5 text-xs font-semibold text-blue-700 transition-colors hover:bg-blue-200"
-        title="View all bookings for this asset"
-      >
-        {row.original.bookingCount}
-      </Link>
-    ),
-  },
-  {
-    accessorKey: "totalDaysBooked",
-    header: "Total Days",
-    cell: ({ row }) => (
-      <span className="text-sm text-gray-700">
-        {row.original.totalDaysBooked}
-      </span>
-    ),
-  },
-  {
-    id: "avgDuration",
-    header: AvgDurationHeader,
-    // Compute average duration for sorting
-    accessorFn: (row) =>
-      row.bookingCount > 0 ? row.totalDaysBooked / row.bookingCount : 0,
-    cell: ({ row }) => {
-      const avgDays = row.getValue("avgDuration") as number;
-
-      // Visual bar capped at 14 days (2 weeks) as reference max
-      const barPercent = Math.min((avgDays / 14) * 100, 100);
-
-      return (
-        <div className="flex items-center gap-2">
-          <div className="h-2 w-12 overflow-hidden rounded-full bg-gray-100">
-            <div
-              className={tw(
-                "h-full rounded-full",
-                avgDays >= 7
-                  ? "bg-blue-600"
-                  : avgDays >= 3
-                  ? "bg-blue-400"
-                  : "bg-blue-200"
-              )}
-              style={{ width: `${barPercent}%` }}
-            />
-          </div>
-          <span className="whitespace-nowrap text-xs text-gray-600">
-            {avgDays >= 1
-              ? `${avgDays.toFixed(1)}d`
-              : `${Math.round(avgDays * 24)}h`}
-          </span>
-        </div>
-      );
-    },
-  },
-  {
-    accessorKey: "category",
-    header: "Category",
-    cell: ({ row }) =>
-      row.original.category || <span className="text-gray-400">—</span>,
-  },
-  {
-    accessorKey: "location",
-    header: "Location",
-    cell: ({ row }) =>
-      row.original.location || <span className="text-gray-400">—</span>,
-  },
-];
-
-function TopBookedAssetsContent({
-  rows,
-  kpis,
-  totalRows,
-  timeframeLabel,
-  topBookedAsset,
-  onRowClick,
-}: {
-  rows: TopBookedAssetRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  timeframeLabel?: string;
-  /** The #1 most booked asset - independent of pagination */
-  topBookedAsset?: TopBookedAssetRow | null;
-  onRowClick?: (row: TopBookedAssetRow) => void;
-}) {
-  // Column definitions for top booked assets table.
-  // Memoized so cell function refs are stable across re-renders — see the
-  // detailed loop explanation in CustodySnapshotContent above. Cells here
-  // are pure (no closure variables), so deps are empty.
-  const columns: ColumnDef<TopBookedAssetRow>[] = useMemo(
-    () => TOP_BOOKED_ASSETS_COLUMNS,
-    []
-  );
-
-  // Extract KPI values
-  const totalBookings =
-    (kpis.find((k) => k.id === "total_bookings")?.rawValue as number) || 0;
-  const uniqueAssetsBooked =
-    (kpis.find((k) => k.id === "unique_assets_booked")?.rawValue as number) ||
-    0;
-  const avgBookingsPerAsset =
-    (kpis.find((k) => k.id === "avg_bookings_per_asset")?.rawValue as number) ||
-    0;
-
-  // Most booked asset comes from server (independent of pagination)
-  const topAsset = topBookedAsset || null;
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span className="text-3xl font-semibold text-gray-900">
-                {totalBookings}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                Total Bookings
-              </span>
-              <span className="text-xs text-gray-500">
-                {uniqueAssetsBooked} assets booked
-              </span>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Avg per Asset</span>
-              <span className="text-lg font-medium text-gray-900">
-                {avgBookingsPerAsset.toFixed(1)}
-              </span>
-            </div>
-
-            {/* Most booked asset with image */}
-            {topAsset && (
-              <div className="flex flex-col">
-                <span className="text-xs text-gray-500">Most Booked</span>
-                <Link
-                  to={`/assets/${topAsset.assetId}`}
-                  className="group -mx-1.5 mt-0.5 flex items-center gap-2 rounded-md px-1.5 py-1 transition-colors hover:bg-gray-50"
-                >
-                  <AssetImage
-                    asset={{
-                      id: topAsset.assetId,
-                      thumbnailImage: topAsset.thumbnailImage,
-                    }}
-                    alt={`Image of ${topAsset.assetName}`}
-                    className="size-8 rounded object-cover ring-1 ring-gray-200"
-                  />
-                  <div className="flex min-w-0 flex-col">
-                    <span
-                      className="max-w-[200px] truncate text-sm font-medium text-gray-900 group-hover:text-primary-600"
-                      title={topAsset.assetName}
-                    >
-                      {topAsset.assetName}
-                    </span>
-                    <span className="text-xs text-gray-500">
-                      {topAsset.bookingCount} booking
-                      {topAsset.bookingCount !== 1 ? "s" : ""} · #1 most booked
-                    </span>
-                  </div>
-                </Link>
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* Context guidance */}
-        <div className="border-t border-gray-100 bg-gray-50/50 px-4 py-2 md:px-6">
-          <p className="text-xs text-gray-500">
-            <span className="font-medium text-gray-600">
-              Understanding this report:
-            </span>{" "}
-            Shows which assets are booked most frequently during{" "}
-            <span className="font-medium">
-              {(timeframeLabel || "the selected period").toLowerCase()}
-            </span>
-            . <span className="italic">Total Days</span> = cumulative booking
-            days. <span className="italic">Avg Duration</span> = typical booking
-            length per checkout. Use longer timeframes (30+ days) for meaningful
-            duration trends.
-          </p>
-        </div>
-      </div>
-
-      {/* Data table — fills remaining vertical space inside the route's flex column
-          and scrolls internally when row count exceeds the visible area. */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">Top Assets</h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          onRowClick={onRowClick}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No booking data"
-              description="No assets have been booked within the selected timeframe."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R10: Asset Distribution Content
-// -----------------------------------------------------------------------------
-
-function AssetDistributionContent({
-  kpis,
-  distributionBreakdown,
-}: {
-  kpis: ReportKpi[];
-  distributionBreakdown?: DistributionBreakdown;
-}) {
-  const navigate = useNavigate();
-
-  // Navigate to assets filtered by the clicked item
-  // IDs match the special filter values: "uncategorized", "without-location", or actual IDs
-  const handleCategoryClick = (item: { id: string }) => {
-    void navigate(`/assets?category=${encodeURIComponent(item.id)}`);
-  };
-
-  const handleLocationClick = (item: { id: string }) => {
-    void navigate(`/assets?location=${encodeURIComponent(item.id)}`);
-  };
-
-  const handleStatusClick = (item: { id: string }) => {
-    void navigate(`/assets?status=${encodeURIComponent(item.id)}`);
-  };
-
-  // Extract KPI values
-  const totalAssets =
-    (kpis.find((k) => k.id === "total_assets")?.rawValue as number) || 0;
-  const totalValue =
-    (kpis.find((k) => k.id === "total_value")?.rawValue as number) || 0;
-  const totalCategories =
-    (kpis.find((k) => k.id === "total_categories")?.rawValue as number) || 0;
-  const totalLocations =
-    (kpis.find((k) => k.id === "total_locations")?.rawValue as number) || 0;
-
-  return (
-    <div className="space-y-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span className="text-3xl font-semibold text-gray-900">
-                {totalAssets}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                Total Assets
-              </span>
-              <span className="text-xs text-gray-500">
-                Across {totalCategories} categories, {totalLocations} locations
-              </span>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Total Value</span>
-              <span className="text-lg font-medium text-gray-900">
-                {totalValue > 0 ? `$${totalValue.toLocaleString()}` : "—"}
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Distribution donut charts - clickable to filter assets */}
-      {distributionBreakdown && (
-        <div className="grid gap-6 lg:grid-cols-3">
-          <DistributionDonut
-            title="By Category"
-            data={distributionBreakdown.byCategory}
-            emptyMessage="No categories defined"
-            maxLegendItems={5}
-            onItemClick={handleCategoryClick}
-          />
-          <DistributionDonut
-            title="By Location"
-            data={distributionBreakdown.byLocation}
-            emptyMessage="No locations defined"
-            maxLegendItems={5}
-            onItemClick={handleLocationClick}
-          />
-          <DistributionDonut
-            title="By Status"
-            data={distributionBreakdown.byStatus}
-            emptyMessage="No status data"
-            maxLegendItems={5}
-            onItemClick={handleStatusClick}
-          />
-        </div>
-      )}
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R1: Asset Inventory Content
-// -----------------------------------------------------------------------------
-
-/**
- * Column definitions for the Asset Inventory table, hoisted to module
- * scope so cell function refs stay stable across re-renders. Same loop
- * fix as IDLE_ASSETS_COLUMNS.
- */
-const ASSET_INVENTORY_COLUMNS: ColumnDef<AssetInventoryRow>[] = [
-  {
-    accessorKey: "assetName",
-    header: "Asset",
-    cell: ({ row }) => (
-      <AssetCell
-        name={row.original.assetName}
-        thumbnailImage={row.original.thumbnailImage}
-        assetId={row.original.assetId}
-      />
-    ),
-  },
-  {
-    accessorKey: "status",
-    header: "Status",
-    cell: ({ row }) => (
-      <StatusCell
-        status={formatAssetStatus(row.original.status)}
-        variant={getAssetStatusVariant(row.original.status)}
-      />
-    ),
-  },
-  {
-    accessorKey: "category",
-    header: "Category",
-    cell: ({ row }) =>
-      row.original.category || <span className="text-gray-400">—</span>,
-  },
-  {
-    accessorKey: "location",
-    header: "Location",
-    cell: ({ row }) =>
-      row.original.location || <span className="text-gray-400">—</span>,
-  },
-  {
-    accessorKey: "custodian",
-    header: "Assigned to",
-    cell: ({ row }) =>
-      row.original.custodian || <span className="text-gray-400">—</span>,
-  },
-  {
-    accessorKey: "valuation",
-    header: "Value",
-    cell: ({ row }) =>
-      row.original.valuation ? (
-        `$${row.original.valuation.toLocaleString()}`
-      ) : (
-        <span className="text-gray-400">—</span>
-      ),
-  },
-  {
-    accessorKey: "createdAt",
-    header: "Created",
-    cell: ({ row }) => <DateCell date={row.original.createdAt} />,
-  },
-];
-
-function AssetInventoryContent({
-  rows,
-  kpis,
-  totalRows,
-  onRowClick,
-}: {
-  rows: AssetInventoryRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  onRowClick?: (row: AssetInventoryRow) => void;
-}) {
-  const columns: ColumnDef<AssetInventoryRow>[] = useMemo(
-    () => ASSET_INVENTORY_COLUMNS,
-    []
-  );
-
-  // Extract KPI values
-  const totalAssets =
-    (kpis.find((k) => k.id === "total_assets")?.rawValue as number) || 0;
-  const totalValue =
-    (kpis.find((k) => k.id === "total_value")?.rawValue as number) || 0;
-  const availableCount =
-    (kpis.find((k) => k.id === "available_count")?.rawValue as number) || 0;
-  const inCustodyCount =
-    (kpis.find((k) => k.id === "in_custody_count")?.rawValue as number) || 0;
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span className="text-3xl font-semibold text-gray-900">
-                {totalAssets}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                Total Assets
-              </span>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Total Value</span>
-              <span className="text-lg font-medium text-gray-900">
-                {totalValue > 0 ? `$${totalValue.toLocaleString()}` : "—"}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Not in use</span>
-              <span className="text-lg font-medium text-gray-900">
-                {availableCount}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Assigned</span>
-              <span className="text-lg font-medium text-gray-900">
-                {inCustodyCount}
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Data table */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">Assets</h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          onRowClick={onRowClick}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No assets"
-              description="Your inventory is empty."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R9: Monthly Booking Trends Content
-// -----------------------------------------------------------------------------
-
-function MonthlyBookingTrendsContent({
-  rows,
-  kpis,
-  totalRows,
-  chartSeries,
-}: {
-  rows: MonthlyBookingTrendRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  chartSeries?: ChartSeries[];
-}) {
-  // Column definitions for trends table
-  const columns: ColumnDef<MonthlyBookingTrendRow>[] = [
-    {
-      accessorKey: "month",
-      header: "Month",
-      cell: ({ row }) => (
-        <span className="font-medium">{row.original.month}</span>
-      ),
-    },
-    {
-      accessorKey: "bookingsCreated",
-      header: "Bookings Created",
-      cell: ({ row }) => <NumberCell value={row.original.bookingsCreated} />,
-    },
-    {
-      accessorKey: "bookingsCompleted",
-      header: "Bookings Completed",
-      cell: ({ row }) => <NumberCell value={row.original.bookingsCompleted} />,
-    },
-    {
-      accessorKey: "momChange",
-      header: "vs Last Month",
-      cell: ({ row }) => {
-        const change = row.original.momChange;
-        if (change === null) return <span className="text-gray-400">—</span>;
-        return (
-          <span
-            className={tw(
-              "inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-semibold",
-              change > 0
-                ? "bg-green-100 text-green-700"
-                : change < 0
-                ? "bg-red-100 text-red-700"
-                : "bg-gray-100 text-gray-700"
-            )}
-          >
-            {change > 0 ? "+" : ""}
-            {change}%
-          </span>
-        );
-      },
-    },
-  ];
-
-  // Extract KPI values
-  const totalBookings =
-    (kpis.find((k) => k.id === "total_bookings")?.rawValue as number) || 0;
-  const avgMonthly =
-    (kpis.find((k) => k.id === "avg_monthly_bookings")?.rawValue as number) ||
-    0;
-  const peakMonth = kpis.find((k) => k.id === "peak_month")?.value || "—";
-  const trendKpi = kpis.find((k) => k.id === "trend_direction");
-  const trendDirection = trendKpi?.value || "Stable";
-  const trendDeltaStr = trendKpi?.delta;
-  const trendDelta = trendDeltaStr ? Number(trendDeltaStr) : null;
-  const trendDescription = trendKpi?.description;
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span className="text-3xl font-semibold text-gray-900">
-                {totalBookings}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                Total Bookings
-              </span>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Average per Month</span>
-              <span className="text-lg font-medium text-gray-900">
-                {avgMonthly}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Peak Month</span>
-              <span className="text-lg font-medium text-gray-900">
-                {peakMonth}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">
-                Trend (Last 2 Months)
-              </span>
-              <div
-                className="group relative cursor-help"
-                title={
-                  trendDescription ||
-                  "Compares the most recent month to the previous month"
-                }
-              >
-                <div className="flex items-center gap-1.5">
-                  <span
-                    className={tw(
-                      "text-lg font-medium",
-                      trendDirection === "Increasing"
-                        ? "text-green-600"
-                        : trendDirection === "Decreasing"
-                        ? "text-red-600"
-                        : "text-gray-900"
-                    )}
-                  >
-                    {trendDirection}
-                  </span>
-                  {trendDelta !== null && (
-                    <span
-                      className={tw(
-                        "text-sm",
-                        trendDelta > 0 ? "text-green-600" : "text-red-600"
-                      )}
-                    >
-                      ({trendDelta > 0 ? "+" : ""}
-                      {trendDelta})
-                    </span>
-                  )}
-                </div>
-                {/* Hover tooltip with more details */}
-                {trendDescription && (
-                  <div className="invisible absolute bottom-full left-1/2 z-10 mb-2 -translate-x-1/2 whitespace-nowrap rounded bg-gray-900 px-2 py-1 text-xs text-white opacity-0 transition-opacity group-hover:visible group-hover:opacity-100">
-                    {trendDescription}
-                    <div className="absolute left-1/2 top-full -translate-x-1/2 border-4 border-transparent border-t-gray-900" />
-                  </div>
-                )}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Bar chart - proper Recharts visualization */}
-      {chartSeries && chartSeries[0]?.data.length > 0 && (
-        <ChartCard title="Booking Volume by Month">
-          <div className="h-64">
-            <BarChart
-              series={chartSeries}
-              radius={4}
-              tooltipFormatter={(value) => `${value} bookings`}
-            />
-          </div>
-        </ChartCard>
-      )}
-
-      {/* Data table */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">
-            Monthly Breakdown
-          </h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No trend data"
-              description="No bookings in the selected timeframe."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R8: Asset Utilization Content
-// -----------------------------------------------------------------------------
-
-/**
- * Column definitions for the Asset Utilization table, hoisted to module
- * scope. Same loop fix as IDLE_ASSETS_COLUMNS.
- */
-const ASSET_UTILIZATION_COLUMNS: ColumnDef<AssetUtilizationRow>[] = [
-  {
-    accessorKey: "assetName",
-    header: "Asset",
-    cell: ({ row }) => (
-      <AssetCell
-        name={row.original.assetName}
-        thumbnailImage={row.original.thumbnailImage}
-        assetId={row.original.assetId}
-      />
-    ),
-  },
-  {
-    accessorKey: "utilizationRate",
-    header: "Usage Rate",
-    cell: ({ row }) => {
-      const rate = row.original.utilizationRate;
-      return (
-        <div className="flex items-center gap-2">
-          <div className="h-2 w-20 overflow-hidden rounded-full bg-gray-200">
-            <div
-              className="h-full rounded-full bg-primary-500"
-              style={{ width: `${Math.min(rate, 100)}%` }}
-            />
-          </div>
-          <span className="text-xs font-semibold tabular-nums text-gray-900">
-            {rate}%
-          </span>
-        </div>
-      );
-    },
-  },
-  {
-    accessorKey: "daysInUse",
-    header: "Days Booked",
-    cell: ({ row }) => (
-      <span>
-        {row.original.daysInUse} / {row.original.totalDays}
-      </span>
-    ),
-  },
-  {
-    accessorKey: "bookingCount",
-    header: "Bookings",
-    cell: ({ row }) => <NumberCell value={row.original.bookingCount} />,
-  },
-  {
-    accessorKey: "category",
-    header: "Category",
-    cell: ({ row }) =>
-      row.original.category || <span className="text-gray-400">—</span>,
-  },
-  {
-    accessorKey: "valuation",
-    header: "Value",
-    cell: ({ row }) =>
-      row.original.valuation ? (
-        `$${row.original.valuation.toLocaleString()}`
-      ) : (
-        <span className="text-gray-400">—</span>
-      ),
-  },
-];
-
-function AssetUtilizationContent({
-  rows,
-  kpis,
-  totalRows,
-  onRowClick,
-}: {
-  rows: AssetUtilizationRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  onRowClick?: (row: AssetUtilizationRow) => void;
-}) {
-  const columns: ColumnDef<AssetUtilizationRow>[] = useMemo(
-    () => ASSET_UTILIZATION_COLUMNS,
-    []
-  );
-
-  // Extract KPI values
-  const avgUtilization =
-    (kpis.find((k) => k.id === "avg_utilization")?.rawValue as number) || 0;
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric with visual bar */}
-          <div className="flex items-center gap-4">
-            <div className="flex flex-col gap-2">
-              <div className="flex items-baseline gap-2">
-                <span className="text-4xl font-bold tabular-nums text-gray-900">
-                  {avgUtilization}%
-                </span>
-                <span className="text-sm font-medium text-gray-500">
-                  average usage
-                </span>
-              </div>
-              <div className="h-2 w-48 overflow-hidden rounded-full bg-gray-200">
-                <div
-                  className="h-full rounded-full bg-primary-500"
-                  style={{ width: `${Math.min(avgUtilization, 100)}%` }}
-                />
-              </div>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Total Assets</span>
-              <span className="text-lg font-medium text-gray-900">
-                {totalRows}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Total Booking Days</span>
-              <span className="text-lg font-medium text-gray-900">
-                {kpis.find((k) => k.id === "total_booking_days")?.value || "0"}
-              </span>
-            </div>
-          </div>
-        </div>
-
-        {/* How utilization is calculated */}
-        <div className="border-t border-gray-100 bg-gray-50 px-4 py-3 md:px-6">
-          <p className="text-xs text-gray-500">
-            <span className="font-medium text-gray-700">
-              How it's calculated:
-            </span>{" "}
-            Usage rate = (total days booked ÷ days in period) × 100. Rates above
-            100% indicate overlapping bookings where the asset was reserved
-            multiple times simultaneously—a sign of high demand.
-          </p>
-        </div>
-      </div>
-
-      {/* Data table */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">
-            Asset Usage Rates
-          </h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          onRowClick={onRowClick}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No usage data"
-              description="No booking activity to calculate usage rates."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
-}
-
-// -----------------------------------------------------------------------------
-// R7: Asset Activity Content
-// -----------------------------------------------------------------------------
-
-/**
- * Column definitions for the Asset Activity table, hoisted to module
- * scope. Same loop fix as IDLE_ASSETS_COLUMNS.
- */
-const ASSET_ACTIVITY_COLUMNS: ColumnDef<AssetActivityRow>[] = [
-  {
-    accessorKey: "assetName",
-    header: "Asset",
-    cell: ({ row }) => (
-      <AssetCell
-        name={row.original.assetName}
-        thumbnailImage={row.original.thumbnailImage}
-        assetId={row.original.assetId}
-      />
-    ),
-  },
-  {
-    accessorKey: "activityType",
-    header: "Activity",
-    cell: ({ row }) => {
-      const type = row.original.activityType;
-      // Plain English labels — no jargon or internal codes
-      const labels: Record<string, string> = {
-        CREATED: "Asset created",
-        UPDATED: "Asset updated",
-        CUSTODY_ASSIGNED: "Assigned to team member",
-        CUSTODY_RELEASED: "Returned from team member",
-        BOOKING_CHECKED_OUT: "Checked out",
-        BOOKING_CHECKED_IN: "Checked in",
-        LOCATION_CHANGED: "Location changed",
-        CATEGORY_CHANGED: "Category changed",
-      };
-      // Semantic colors for activity types
-      const colors: Record<string, string> = {
-        CREATED: "bg-green-100 text-green-700",
-        UPDATED: "bg-blue-100 text-blue-700",
-        CUSTODY_ASSIGNED: "bg-violet-100 text-violet-700",
-        CUSTODY_RELEASED: "bg-violet-100 text-violet-700",
-        BOOKING_CHECKED_OUT: "bg-orange-100 text-orange-700",
-        BOOKING_CHECKED_IN: "bg-green-100 text-green-700",
-        LOCATION_CHANGED: "bg-blue-100 text-blue-700",
-        CATEGORY_CHANGED: "bg-blue-100 text-blue-700",
-      };
-      return (
-        <span
-          className={tw(
-            "inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-xs font-medium",
-            colors[type] || "bg-gray-100 text-gray-700"
-          )}
-        >
-          {labels[type] || type}
-        </span>
-      );
-    },
-  },
-  {
-    accessorKey: "performedBy",
-    header: "Changed by",
-    cell: ({ row }) =>
-      row.original.performedBy || <span className="text-gray-400">System</span>,
-  },
-  {
-    accessorKey: "occurredAt",
-    header: "Date & Time",
-    cell: ({ row }) => <DateCell date={row.original.occurredAt} />,
-  },
-];
-
-function AssetActivityContent({
-  rows,
-  kpis,
-  totalRows,
-  onRowClick,
-}: {
-  rows: AssetActivityRow[];
-  kpis: ReportKpi[];
-  totalRows: number;
-  onRowClick?: (row: AssetActivityRow) => void;
-}) {
-  const columns: ColumnDef<AssetActivityRow>[] = useMemo(
-    () => ASSET_ACTIVITY_COLUMNS,
-    []
-  );
-
-  // Extract KPI values
-  const totalActivities =
-    (kpis.find((k) => k.id === "total_activities")?.rawValue as number) || 0;
-  const custodyChanges =
-    (kpis.find((k) => k.id === "custody_changes")?.rawValue as number) || 0;
-  const bookingActivities =
-    (kpis.find((k) => k.id === "booking_activities")?.rawValue as number) || 0;
-  const mostActiveAsset =
-    kpis.find((k) => k.id === "most_active_asset")?.value || "—";
-
-  return (
-    <div className="flex flex-col gap-4">
-      {/* Hero section */}
-      <div className="rounded border border-gray-200 bg-white">
-        <div className="flex flex-col gap-4 p-4 md:flex-row md:items-center md:justify-between md:p-6">
-          {/* Main metric */}
-          <div className="flex items-center gap-4">
-            <div className="flex items-baseline gap-1">
-              <span className="text-3xl font-semibold text-gray-900">
-                {totalActivities}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-sm font-medium text-gray-700">
-                Total Activities
-              </span>
-            </div>
-          </div>
-
-          {/* Supporting stats */}
-          <div className="flex gap-6 border-t border-gray-100 pt-3 md:border-l md:border-t-0 md:pl-6 md:pt-0">
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Assignments</span>
-              <span className="text-lg font-medium text-gray-900">
-                {custodyChanges}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Check-ins/outs</span>
-              <span className="text-lg font-medium text-gray-900">
-                {bookingActivities}
-              </span>
-            </div>
-            <div className="flex flex-col">
-              <span className="text-xs text-gray-500">Most Active Asset</span>
-              <span
-                className="max-w-[120px] truncate text-lg font-medium text-gray-900"
-                title={mostActiveAsset}
-              >
-                {mostActiveAsset}
-              </span>
-            </div>
-          </div>
-        </div>
-
-        {/* Activity types legend */}
-        <div className="border-t border-gray-100 bg-gray-50 px-4 py-3 md:px-6">
-          <div className="flex flex-wrap items-center gap-x-4 gap-y-2">
-            <span className="text-xs font-medium text-gray-700">
-              Activity types:
-            </span>
-            <span className="inline-flex items-center gap-1 rounded-full bg-green-100 px-2 py-0.5 text-xs font-medium text-green-700">
-              Asset created
-            </span>
-            <span className="inline-flex items-center gap-1 rounded-full bg-blue-100 px-2 py-0.5 text-xs font-medium text-blue-700">
-              Updated / Moved
-            </span>
-            <span className="inline-flex items-center gap-1 rounded-full bg-violet-100 px-2 py-0.5 text-xs font-medium text-violet-700">
-              Custody changed
-            </span>
-            <span className="inline-flex items-center gap-1 rounded-full bg-orange-100 px-2 py-0.5 text-xs font-medium text-orange-700">
-              Checked out
-            </span>
-            <span className="inline-flex items-center gap-1 rounded-full bg-green-100 px-2 py-0.5 text-xs font-medium text-green-700">
-              Checked in
-            </span>
-          </div>
-        </div>
-      </div>
-
-      {/* Data table */}
-      <div className="overflow-hidden rounded border border-gray-200 bg-white">
-        <div className="flex items-center gap-2 border-b border-gray-100 px-4 py-3 md:px-6">
-          <h3 className="text-sm font-semibold text-gray-900">
-            Recent Activity
-          </h3>
-          <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-600">
-            {totalRows}
-          </span>
-        </div>
-        <ReportTable
-          data={rows}
-          columns={columns}
-          onRowClick={onRowClick}
-          emptyContent={
-            <ReportEmptyState
-              reason="no_data"
-              title="No activity"
-              description="No asset activity recorded in this timeframe."
-            />
-          }
-        />
-      </div>
-    </div>
-  );
 }
